requirements.txt
annotated-types==0.7.0
beautifulsoup4==4.14.3
brotli==1.2.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
et-xmlfile==2.0.0
idna==3.11
markdown-it-py==4.0.0
mdurl==0.1.2
openpyxl==3.1.5
prompt-toolkit==3.0.52
pydantic==2.12.5
pydantic-core==2.41.5
pygments==2.19.2
questionary==2.1.1
requests==2.32.5
rich==14.2.0
soupsieve==2.8
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.5.0
wcwidth==0.2.14
datetime==5.2.0